// src/application/enums/tokens/job-token.enum.ts
export enum JOB_TOKEN {
	JOB_REPOSITORY = "JOB_REPOSITORY",
	JOB_USE_CASE = "JOB_USE_CASE",
	CREATE_JOB_USE_CASE = "CREATE_JOB_USE_CASE",
	GET_RECRUITER_JOBS_USE_CASE = "GET_RECRUITER_JOBS_USE_CASE",
	APPLY_JOB_USE_CASE = "APPLY_JOB_USE_CASE",
	GET_ALL_JOBS_USE_CASE = "GET_ALL_JOBS_USE_CASE",
	BLOCK_UNBLOCK_JOB_USE_CASE = "BLOCK_UNBLOCK_JOB_USE_CASE",
	UPDATE_JOB_USE_CASE = "UPDATE_JOB_USE_CASE",
	GET_CANDIDATE_JOBS_USE_CASE = "GET_CANDIDATE_JOBS_USE_CASE",
	GET_JOB_BY_ID_USE_CASE = "GET_JOB_BY_ID_USE_CASE",
	JOB_APPLICATION_REPOSITORY = "JOB_APPLICATION_REPOSITORY",
}